resolve conflict between the 2 python migration viewmodels
